Recreate /page/index
