+++ assets/sass/app.scss
@@ -1,2 +1,27 @@
 @import 'variables';
 @import 'common';
+
+html {
+  font-size: 16px;
+}
+
+#app {
+  background: linear-gradient(to bottom, #ff3e94, #ff655d);
+}
+
+* {
+  box-sizing: border-box;
+  font-family: 'Hiragino Kaku Gothic Pro', sans-serif;
+  color: white;
+}
+
+.container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+}
+
+a {
+  text-decoration: none;
+}
