+++ assets/sass/common.scss
@@ -1,96 +1,136 @@
 @import 'variables';
 
-* {
-  box-sizing: border-box;
-  font-family: 'Hiragino Kaku Gothic Pro', sans-serif;
-}
-body {
-  background-color: $color-bg;
-}
-a {
-  text-decoration: none;
-}
-.container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-$primary-color: #fe5492;
-
-.shadow{
-  box-shadow: none;
-}
-
-.radio-tile-group {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  width: 100%;
-
-  //set dimensions for invisible container
-  .input-container {
-    position: relative;
-    height: 5rem;
-    width: 5rem;
-    margin: 0.8rem 0.2rem;
-
-    //make actual radio input invisible
-    // + stretch to fill container
-    .radio-button {
-      opacity: 0;
-      position: absolute;
-      top: 0;
-      left: 0;
-      height: 100%;
-      width: 100%;
-      margin: 0;
-      cursor: pointer;
-    }
-
-    //default tile styles
-    .radio-tile {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      justify-content: center;
-      width: 100%;
-      height: 100%;
-      border: 2px solid $primary-color;
-      border-radius: 5px;
-      padding: 1rem;
-      transition: transform 300ms ease;
-    }
-    .icon svg {
-      fill: $primary-color;
-      width: 2rem;
-      height: 2rem;
-    }
-    .radio-tile-label {
-      text-align: center;
-      font-size: 0.5rem;
-      font-weight: 600;
-      text-transform: uppercase;
-      letter-spacing: 1px;
-      color: $primary-color;
-    }
-
-    //active tile styles
-    .radio-button:checked + .radio-tile {
-      background-color: $primary-color;
-      border: 2px solid $primary-color;
-      color: white;
-      transform: scale(1.1, 1.1);
-
-      .icon svg {
-        fill: white;
-        background-color: $primary-color;
-      }
-      .radio-tile-label {
-        color: white;
-        background-color: $primary-color;
-      }
-    }
+@media all and (min-width: 769px) {
+  .q-phrase {
+    padding: 200px 0 60px;
+  }
+  .skip {
+    padding-top: 60px;
+  }
+  .radio-tile-group input {
+    display: none;
+  }
+  .radio-label {
+    border: 1px solid $font-color;
+    border-radius: 5px;
+    padding: 2.5rem 2.5rem;
+    margin: 0 1rem;
+    background: rgba(255, 255, 255, 0.3);
+  }
+  .scroll {
+    overflow-y: scroll;
   }
 }
 
+@media all and (max-width: 768px) {
+  .q-phrase {
+    padding: 100px 0 50px;
+  }
+  .skip {
+    padding-top: 40px;
+  }
+  .radio-tile-group input {
+    display: none;
+  }
+  .radio-label {
+    border: 1px solid $font-color;
+    border-radius: 5px;
+    padding: 20px;
+    margin: 0 1rem;
+    background: rgba(255, 255, 255, 0.3);
+  }
+  .scroll {
+    overflow-y: scroll;
+  }
+}
+// 以下過去の遺物
+// .plan-img {
+//   width: 400px;
+//   height: 400px;
+//   object-fit: contain;
+// }
+// .thumb {
+//   max-width: 250px;
+//   max-height: 250px;
+//   width: 100%;
+//   height: 100%;
+//   object-fit: cover;
+// }
+
+// $primary-color: #fe5492;
+
+// .shadow{
+//   box-shadow: none;
+// }
+
+// .radio-tile-group {
+//   display: flex;
+//   flex-wrap: wrap;
+//   justify-content: center;
+//   width: 100%;
+
+//   //set dimensions for invisible container
+//   .input-container {
+//     position: relative;
+//     height: 5rem;
+//     width: 5rem;
+//     margin: 0.8rem 0.2rem;
+
+//     //make actual radio input invisible
+//     // + stretch to fill container
+//     .radio-button {
+//       opacity: 0;
+//       position: absolute;
+//       top: 0;
+//       left: 0;
+//       height: 100%;
+//       width: 100%;
+//       margin: 0;
+//       cursor: pointer;
+//     }
+
+//     //default tile styles
+//     .radio-tile {
+//       display: flex;
+//       flex-direction: column;
+//       align-items: center;
+//       justify-content: center;
+//       width: 100%;
+//       height: 100%;
+//       border: 2px solid $primary-color;
+//       border-radius: 5px;
+//       padding: 1rem;
+//       transition: transform 300ms ease;
+//     }
+//     .icon svg {
+//       fill: $primary-color;
+//       width: 2rem;
+//       height: 2rem;
+//     }
+//     .radio-tile-label {
+//       text-align: center;
+//       font-size: 0.5rem;
+//       font-weight: 600;
+//       text-transform: uppercase;
+//       letter-spacing: 1px;
+//       color: $primary-color;
+//     }
+
+//     //active tile styles
+//     .radio-button:checked + .radio-tile {
+//       background-color: $primary-color;
+//       border: 2px solid $primary-color;
+//       color: white;
+//       transform: scale(1.1, 1.1);
+
+//       .icon svg {
+//         fill: white;
+//         background-color: $primary-color;
+//       }
+//       .radio-tile-label {
+//         color: white;
+//         background-color: $primary-color;
+//       }
+//     }
+//   }
+// }
