+++ assets/sass/phone.scss
@@ -1,78 +1,3 @@
 @media screen and (max-width: 768px) {
-  .radio-tile-group {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    width: 100%;
-
-    //set dimensions for invisible container
-    .input-container {
-      position: relative;
-      height: 5em;
-      width: 5em;
-      margin: 8px;
-
-      //make actual radio input invisible
-      // + stretch to fill container
-      .radio-button {
-        opacity: 0;
-        position: absolute;
-        top: 0;
-        left: 0;
-        height: 100%;
-        width: 100%;
-        margin: 0;
-        cursor: pointer;
-      }
-
-      //default tile styles
-      .radio-tile {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: center;
-        width: 100%;
-        height: 100%;
-        border: 2px solid $primary-color;
-        border-radius: 5px;
-        padding: 1rem;
-        transition: transform 300ms ease;
-      }
-      .icon svg {
-        fill: $primary-color;
-        width: 2rem;
-        height: 2rem;
-      }
-      .radio-tile-label {
-        text-align: center;
-        font-size: 0.5rem;
-        font-weight: 600;
-        text-transform: uppercase;
-        letter-spacing: 1px;
-        color: $primary-color;
-      }
-
-      //active tile styles
-      .radio-button:checked + .radio-tile {
-        background-color: $primary-color;
-        border: 2px solid $primary-color;
-        color: white;
-        transform: scale(1.1, 1.1);
-
-        .icon svg {
-          fill: white;
-          background-color: $primary-color;
-        }
-        .radio-tile-label {
-          color: white;
-          background-color: $primary-color;
-        }
-      }
-    }
-  }
-
-  .wrapper {
-    width: 768px;
-    max-width: 100%;
-  }
+  
 }
