+++ assets/sass/variables.scss
@@ -1,1 +1,1 @@
-$color-bg: white;
+$font-color: white;
